Big changes to about, gallery and services page and overall code
Utilized advice from teacher to create a text wrapper to push all of the text of the edge of the screen improving the aesthetics of the website. Redid all of the layouts of the about, gallery, and services page as a result of research of other outcomes. Overall improved code health by rewriting most of the code to fix (organize, issues, inconsistencies) due to receiving lots of errors from validating my code.