Gallery and images updates
Finished the gallery page adding the last of the images. deleting all of the images not being used in the images folder and putting the images through a file compressor to optimise the images. The all of the images are a lot smaller in size now and as a result the website runs smoother as there is no more unnecessary storage being utilised.